Refactor brokers to split responsibilities more efficiently
Split unified broker into a "metadata" part and an "artifact" part.

The metadata part is responsible /only/ for metadata transformation
(converting Plugin meta.yaml into Che Plugins and adding required
fields). It's designed to be relatively workspace-agnostic; currently
its place is as a separate deployment to resolve Che tooling, but it
could be integrated as a library in other functionality.

The artifact part is responsible /only/ for downloading plugin
extensions given a list of plugin metas. It is designed to run as an
init container on all Theia workspaces.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>