Modify broker.ProcessPlugins method to return data
The plugin brokers should be able to be used as a library in other Go
code, to allow plugin brokering without a full execution of the brokers
code. This change modifies the ProcessPlugin method to return ChePlugins
rather than adding them to the internal storage struct so that results
can be obtained in a more fine-grained manner.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>